Python/examples/protobuf_test_1/requirements.txt
protobuf
protobuf_to_dict
person_pb2
google.protobuf
